Corrected product deletion and storage product addition

+++ zlg_website/views/delete_products.py
@@ -8,6 +8,10 @@ from . import views
 def delete_product(product_id):
     product = Product.query.get_or_404(product_id)
 
+    if product.store_items:
+        flash("Неможливо видалити товар, оскільки він є в наявності.", 'error')
+        return redirect(url_for('views.products'))
+
     db.session.delete(product)
     db.session.commit()
     flash("Товар видалено", 'success')
